rearranging Query Engine classes
