+++ hyperon_das/das.py
@@ -12,7 +12,8 @@ from hyperon_das.exceptions import (
     InvalidQueryEngine,
 )
 from hyperon_das.logger import logger
-from hyperon_das.query_engines import LocalQueryEngine, RemoteQueryEngine
+from hyperon_das.query_engines.local_query_engine import LocalQueryEngine
+from hyperon_das.query_engines.remote_query_engine import RemoteQueryEngine
 from hyperon_das.traverse_engines import TraverseEngine
 from hyperon_das.type_alias import Query
 from hyperon_das.utils import QueryAnswer, get_package_version
